+++ src/lib.rs
@@ -71,122 +71,9 @@
 //! }
 //! ```
 
-use tokio::sync::{
-    mpsc::{
-        unbounded_channel,
-        UnboundedSender,
-        UnboundedReceiver,
-    },
-    oneshot,
-};
-use std::fmt::Debug;
+mod actor;
+mod link;
+mod process;
 
-enum Call<State, Reply> {
-    Ref(fn(&State)),
-    RefMut(fn(&mut State)),
-    RefReply(fn(&State, ReplySender<Reply>), ReplySender<Reply>),
-    RefMutReply(fn(&State, ReplySender<Reply>), ReplySender<Reply>),
-}
-
-/// `Process` holds the `Actor` state and sequentially processes
-/// calls sent from the `Actor`.
-pub struct Process<State, Reply>
-{
-    state: State,
-    receiver: CallReceiver<State, Reply>,
-}
-
-/// `Actor` maintains a connection to its `Process` to allow
-/// `Actor` methods to be implemented via functions sent to
-/// the `Process`.
-pub struct Actor<State, Reply> {
-    sender: CallSender<State, Reply>
-}
-
-impl<State, Reply> Clone for Actor<State, Reply> {
-    /// `Actor` may be freely cloned and share the backing `Process`.
-    fn clone(&self) -> Self {
-        Self {
-            sender: self.sender.clone(),
-        }
-    }
-}
-
-type CallReceiver<State, Reply> = UnboundedReceiver<Call<State, Reply>>;
-type CallSender<State, Reply> = UnboundedSender<Call<State, Reply>>;
-type ReplySender<Reply> = oneshot::Sender<Reply>;
-
-impl<State, Reply> Process<State, Reply>
-    where
-        State: Debug,
-{
-    /// Creates a new (`Process`, `Actor`) pair given an initial state.
-    pub fn new_with_state(state: State) -> (Self, Actor<State, Reply>) {
-        let (sender, receiver) = unbounded_channel();
-        (
-            Self {
-                state,
-                receiver,
-            },
-            Actor::new_with_sender(sender)
-        )
-    }
-
-    /// Starts the `Process`.  The corresponding `Actor` is alive as long
-    /// as the `Process` is running.
-    pub async fn start(&mut self) {
-        while let Some(call) = self.receiver.recv().await {
-            match call {
-                Call::Ref(caller) => {
-                    caller(&self.state);
-                },
-                Call::RefMut(caller) => {
-                    caller(&mut self.state);
-                },
-                Call::RefReply(caller, reply_sender) => {
-                    caller(&self.state, reply_sender);
-                },
-                Call::RefMutReply(caller, reply_sender) => {
-                    caller(&mut self.state, reply_sender);
-                }
-            }
-        }
-    }
-}
-
-impl<State, Reply> Actor<State, Reply>
-    where
-        State: Debug,
-{
-    fn new_with_sender(sender: CallSender<State, Reply>) -> Self {
-        Self {
-            sender,
-        }
-    }
-
-    /// Executes a function in the `Process` with a reference to the state.
-    pub fn call_ref(&self, caller: fn(&State)) {
-        self.sender.send(Call::Ref(caller)).ok();
-    }
-
-    /// Executes a function in the `Process` with a mutable reference to the state.
-    pub fn call_ref_mut(&self, caller: fn(&mut State)) {
-        self.sender.send(Call::RefMut(caller)).ok();
-    }
-
-    /// Executes a function in the `Process` with a reference to the state and a `ReplySender`.
-    pub async fn call_ref_reply(&self, caller: fn(&State, ReplySender<Reply>)) -> Reply {
-        let (reply_sender, reply_receiver) = oneshot::channel();
-        self.sender.send(Call::RefReply(caller, reply_sender)).ok();
-
-        reply_receiver.await.ok().unwrap()
-    }
-
-    /// Executes a function in the `Process` with a mutable reference to the state and a `ReplySender`.
-    pub async fn call_ref_mut_reply(&self, caller: fn(&State, ReplySender<Reply>)) -> Reply {
-        let (reply_sender, reply_receiver) = oneshot::channel();
-        self.sender.send(Call::RefMutReply(caller, reply_sender)).ok();
-
-        reply_receiver.await.ok().unwrap()
-    }
-}
+pub use actor::Actor;
+pub use process::Process;
